Heavily tweaked idle/run state logic;
This should improve animations.
Added a force to 'help' trella go up slopes, able to increase player
friction as a result.  This should lead to better responsiveness when
dealing with slopes.
Temporarily disabled slope angling, as it is not finished.
Tweaked run animation speed.
Remove velocity slowdown when changing direction.